Lazily generate deletion tables
Previously a very high computational cost was paid to generate deletion tables
when the dictionary was initialized. This modifies the mechanics so that instead
of creating the full table upon load, we create a two-level-deep trie using
the assumption that the user always enters the first character of the word
correctly (this is highly probable and is a mechanic used by other fuzzy
searching implementations). Then, we load the subsections of the trie that
the a suggestion touches upon request.

Overall this gives us around a 20x speed increase when generating the
dictionary initially (from 10 seconds to 500ms), at the cost of a ~150ms
computation whenever we need to convert a subtree.